refactor(cli): split cli.ts (2938 lines) into 7 focused modules
Pure move + re-import — same symbols, no logic changes:
- cli/repl.ts        the interactive REPL (repl, initReplSession, approvals)
- cli/model-setup.ts provider config/factory, /model, context-window probe
- cli/logging.ts     spinner, terminal Reporter, --log ledger, log paths
                     (module state now behind setInteractiveStream())
- cli/banner.ts      welcome banner, startup hint, plan chip, resume replay
- cli/gate-setup.ts  gate resolution (resumed > --accept > --web > auto);
                     BROWSER_CHECK now reused from gate/tool-paths
- cli/repl-commands.ts  /sessions /map /review /trace + metrics line
- cli/web-setup.ts   web scaffold + deps install progress

cli.ts keeps main(), runOnce(), and the one-shot modes (559 lines, was
2938). External import paths preserved via re-exports (providerConfig,
isApproval/isPlanApproval, spinner). Parity test now reads cli/repl.ts.

Full validate green (1890 pass + 5 PTY suites); live REPL boot smoke-tested
on the real binary.
