+++ packages/core/src/cli.ts
@@ -1,35 +1,9 @@
 #!/usr/bin/env bun
 import { join, isAbsolute } from "node:path";
-import { mkdirSync, existsSync } from "node:fs";
-import { readdir } from "node:fs/promises";
-import { Writable } from "node:stream";
-import { createInterface } from "node:readline/promises";
-import { emitKeypressEvents } from "node:readline";
-import { formatHelp, takesArg } from "./cli/commands";
-import { resolveInitialPlanMode } from "./cli/plan-default";
-import { modeById, nextMode } from "./cli/modes";
-import { runConfigMenu } from "./cli/config-menu";
-import { runCapabilityMenu } from "./cli/capability-menu";
-import { openScaffoldInRepl } from "./cli/repl-scaffold";
-import { openRecipePicker } from "./cli/repl-recipe";
-import { pickCommand, type IPaletteView } from "./render/command-menu";
-import {
-  pickFileInline,
-  filterFiles,
-  formatCompletionRows,
-  shouldOpenAtPicker,
-  type IPickerView,
-} from "./render/file-menu";
-import { listWorkspaceFiles, readFiles, runShellCommand } from "./lib/fs";
 import { renderCheck } from "./browser";
-import { composeMessage } from "./loop/prompt";
 import {
   runTask,
   RUN_STATUS,
-  Session,
-  LedgerWriter,
-  ledgerTypeFor,
-  PLAN_APPROVED_NOTE,
   reviewChange,
   reviewRepair,
   formatReport,
@@ -41,11 +15,8 @@ import {
   type IFeature,
   type IGreenfieldDeps,
   type Reporter,
-  type SetupWebFn,
 } from "./loop";
 import { modelAgent } from "./agent";
-import { buildAndPersistMap, mapStatus, forgetMap } from "./codebase";
-import { parseEventLog, formatTrace } from "./eval";
 import { loadRecipes, findRecipe } from "./config/recipes";
 import {
   parseArgs,
@@ -53,80 +24,19 @@ import {
   isOneShot,
   scopeOf,
   cliUsage,
-  WHOLE_REPO,
   type ICliArgs,
 } from "./cli/args";
-import { makeSpinner, spinnerPhase } from "./render/spinner";
 import { validate } from "./validate";
-import { isPolicyMode } from "./policy";
-import { startEditor, type IEditorHandle } from "./editor";
-import { renderEditor } from "./editor/view";
-import { flags } from "./config/flags";
-import {
-  PROVIDER_LIMITS,
-  PROVIDER_DEFAULTS,
-  OpenAICompatibleProvider,
-  type IOpenAICompatibleConfig,
-} from "./inference";
-import {
-  resolveActiveModel,
-  resolveModelByName,
-  setActiveModel,
-  loadModelsConfig,
-  resolveApiKey,
-  type IModelEntry,
-} from "./models-config";
-import {
-  renderEvent,
-  renderMessage,
-  renderStatus,
-  userBubble,
-  agentCardTop,
-  agentCardBottom,
-  agentBar,
-  makeAgentRail,
-  StatusBar,
-  MIN_ROWS,
-  welcomeBanner,
-  STYLE,
-  paint,
-  PROMPT_COLS,
-  type IStatusInfo,
-} from "./render";
+import type { OpenAICompatibleProvider } from "./inference";
+import { resolveActiveModel, resolveModelByName } from "./models-config";
 import type { ITask } from "./spec";
-import { loadLedger, activeRules, forgetMemory } from "./loop/memory";
-import {
-  buildGate,
-  buildWebGate,
-  buildWebFix,
-  buildCoreFix,
-  buildWebTypeGate,
-  buildWebTscCheck,
-  makeFileLinter,
-  WEB_PACKS,
-  type FileLinter,
-} from "./gate";
-import {
-  scaffoldWeb,
-  installWebDeps,
-  webGuidance,
-} from "./scaffold/web-scaffold";
-import type { WebFramework } from "./web-templates";
-import { isRecord } from "./lib/guards";
-import {
-  saveSession,
-  latestSession,
-  loadSession,
-  listSessions,
-  pruneSessions,
-  logsDir,
-  type ISessionRecord,
-} from "./session-store";
-import {
-  currentVersion,
-  getUpdateNotice,
-  refreshUpdateCacheInBackground,
-} from "./update-check";
+import { readFiles, runShellCommand } from "./lib/fs";
+import { currentVersion } from "./update-check";
+import { repl } from "./cli/repl";
+import { runMapCommand, runTraceCommand } from "./cli/repl-commands";
+import { makeProvider, modelForRun, envNumber } from "./cli/model-setup";
+import { makeReporter, resolveLogPath } from "./cli/logging";
+import { resolveGate } from "./cli/gate-setup";
 
 /**
  * The tsforge CLI — the product surface over the same engine the eval harness
@@ -153,385 +63,9 @@ import {
  */
 export { parseArgs, applyRecipe, isOneShot, type ICliArgs } from "./cli/args";
 
-/** A unique-enough id for a new session (time + a little randomness). */
-function newSessionId(): string {
-  return `${Date.now().toString(36)}-${Math.random().toString(36).slice(2, 8)}`;
-}
-
-/** Human label for an editable scope (the whole-repo default reads nicer). */
-function scopeLabel(files: string[]): string {
-  return files.length === 1 && files[0] === "**/*"
-    ? "entire workspace"
-    : files.join(", ");
-}
-
-/** The host:port of an API base URL, for the banner (falls back to the raw url). */
-function hostOf(baseUrl: string): string {
-  try {
-    return new URL(baseUrl).host;
-  } catch {
-    return baseUrl;
-  }
-}
-
-/** The active model id + endpoint host, from a wire-config (provider.config) or a
- *  registry entry — both carry `model` + `baseUrl`. */
-function modelInfo(src: { model: string; baseUrl: string }): {
-  model: string;
-  endpoint: string;
-} {
-  return { model: src.model, endpoint: hostOf(src.baseUrl) };
-}
-
-/** The model's real context window, read from the server's `/models`
- *  (`max_model_len` — vLLM/OpenAI-compatible). Best-effort: undefined if the
- *  endpoint is unreachable or doesn't report it (caller falls back). 3s cap so a
- *  dead endpoint can't stall CLI startup. */
-async function detectContextWindow(
-  entry: IModelEntry
-): Promise<number | undefined> {
-  const headers: Record<string, string> = {};
-  const key = resolveApiKey(entry);
-
-  if (key !== undefined) {
-    headers.authorization = `Bearer ${key}`;
-  }
-
-  try {
-    const res = await fetch(`${entry.baseUrl}/models`, {
-      headers,
-      signal: AbortSignal.timeout(3000),
-    });
-
-    if (!res.ok) {
-      return undefined;
-    }
-
-    const data: unknown = await res.json();
-
-    if (!isRecord(data) || !Array.isArray(data.data)) {
-      return undefined;
-    }
-
-    const entries = data.data.filter(isRecord);
-    const match = entries.find((e) => e.id === entry.model) ?? entries[0];
-    // vLLM uses `max_model_len`; other servers expose `context_window` or
-    // `max_position_embeddings` — accept whichever is present.
-    const len =
-      match?.max_model_len ??
-      match?.context_window ??
-      match?.max_position_embeddings;
-
-    return typeof len === "number" && Number.isFinite(len) ? len : undefined;
-  } catch {
-    return undefined;
-  }
-}
-
-function frameworkLabel(framework: WebFramework): string {
-  return framework === "react"
-    ? "Vite + React + shadcn/ui + TanStack"
-    : "Vite + TypeScript + Tailwind";
-}
-
-/** The `/metrics` turns-to-green line (loop-efficiency: turns the last green run
- *  took). Extracted so the command switch stays a flat dispatch. */
-function turnsToGreenLine(turns: number | null): string {
-  return turns === null
-    ? "  turns to green: — (no green run yet)\n"
-    : `  turns to green (last): ${String(turns)}\n`;
-}
-
-/** Lay down a stack's skeleton and install its dependencies, reporting progress —
- *  the model can't build until deps resolve. Returns the files actually written and
- *  whether install succeeded so the `scaffold_web` tool can account for the mutation
- *  and tell the model the truth (instead of always claiming "deps installed"). */
-async function setUpWebProject(
-  dir: string,
-  framework: WebFramework,
-  options: { signal?: AbortSignal } = {}
-): Promise<{ files: readonly string[]; depsInstalled: boolean }> {
-  const files = await scaffoldWeb(dir, framework);
-
-  process.stdout.write(`  ↳ installing ${frameworkLabel(framework)}…\n`);
-
-  const depsInstalled = await installWebDeps(dir, options);
-
-  process.stdout.write(
-    depsInstalled
-      ? "  ↳ dependencies ready\n"
-      : "  ⚠ dependency install failed — run `bun install` yourself\n"
-  );
-
-  return { files, depsInstalled };
-}
-
-/** Parse a numeric env var, returning undefined for unset/blank/non-numeric
- *  input (never NaN — a NaN reaching the provider serializes to `null` in the
- *  request body and the model request fails confusingly). */
-function envNumber(name: string): number | undefined {
-  const raw = process.env[name];
-
-  if (raw === undefined || raw.trim().length === 0) {
-    return undefined;
-  }
-
-  const value = Number(raw);
-
-  return Number.isFinite(value) ? value : undefined;
-}
-
-/** Wire-config from a registry entry: API key resolved at use time (inline or
- *  via apiKeyEnv); env still tunes maxTokens/penalty. Shared by initial
- *  construction, `/model` hot-swap, and the interactive eval script — so they
- *  all behave identically. */
-export function providerConfig(entry: IModelEntry): IOpenAICompatibleConfig {
-  const repetitionPenalty = envNumber("TSFORGE_REPETITION_PENALTY");
-
-  return {
-    baseUrl: entry.baseUrl,
-    model: entry.model,
-    apiKey: resolveApiKey(entry),
-    maxTokens:
-      entry.maxTokens ??
-      envNumber("TSFORGE_MAX_TOKENS") ??
-      PROVIDER_LIMITS.maxTokens,
-    // OFF by default: a global repetition penalty also penalizes the rigid,
-    // repetitive tool-call JSON tokens, which pushes the model to NARRATE
-    // instead of emitting tool calls (→ no files written). The StreamGuard is
-    // the targeted loop protection. Opt in only to experiment.
-    ...(repetitionPenalty === undefined ? {} : { repetitionPenalty }),
-    // Provider dialect + escape hatches — passed straight through so any
-    // OpenAI-ish endpoint (DeepSeek, OpenAI o-series, custom gateways) works.
-    ...(entry.reasoning === undefined ? {} : { reasoning: entry.reasoning }),
-    ...(entry.reasoningEffort === undefined
-      ? {}
-      : { reasoningEffort: entry.reasoningEffort }),
-    // Optional override only — guided decoding is auto-detected by endpoint
-    // (local on, DeepSeek cloud off). Passed through when a model entry sets it.
-    ...(entry.guidedDecoding === undefined
-      ? {}
-      : { guidedDecoding: entry.guidedDecoding }),
-    ...(entry.extraBody === undefined ? {} : { extraBody: entry.extraBody }),
-    ...(entry.extraHeaders === undefined
-      ? {}
-      : { extraHeaders: entry.extraHeaders }),
-  };
-}
-
-function makeProvider(entry: IModelEntry): OpenAICompatibleProvider {
-  return new OpenAICompatibleProvider(providerConfig(entry));
-}
-
-/** Catch the common footgun: a cloud baseUrl paired with the leftover qwen
- *  default `model`, which then 400s ("model not supported") on that host. */
-function warnDefaultModelOnRemote(entry: IModelEntry): void {
-  let host: string;
-
-  try {
-    host = new URL(entry.baseUrl).hostname;
-  } catch {
-    return;
-  }
-
-  const remote = host !== "localhost" && host !== "127.0.0.1" && host !== "::1";
-
-  if (remote && entry.model === PROVIDER_DEFAULTS.model) {
-    process.stdout.write(
-      `  ⚠ models.json: model is still "${PROVIDER_DEFAULTS.model}" (the default) but baseUrl is ${host} — set the entry's "model" to a name that host supports.\n`
-    );
-  }
-}
-
-/** Print the model registry with ★ on the active one (the `/model` listing). */
-async function listModels(
-  provider: OpenAICompatibleProvider,
-  activeName: string
-): Promise<void> {
-  const cfg = await loadModelsConfig();
-  const current = modelInfo(provider.config);
-
-  process.stdout.write(
-    `  active: ${activeName} — ${current.model} @ ${current.endpoint}\n`
-  );
-
-  for (const [name, e] of Object.entries(cfg.models)) {
-    const mark = name === activeName ? "★" : " ";
-
-    process.stdout.write(
-      `  ${mark} ${name}  ${e.model} @ ${hostOf(e.baseUrl)}\n`
-    );
-  }
-
-  if (activeName === "env") {
-    process.stdout.write(
-      "  (TSFORGE_* env is overriding the registry — unset it to use /model)\n"
-    );
-  }
-
-  process.stdout.write("  switch with: /model <name>\n");
-}
-
-/** Handle `/model [name]`: no arg lists the registry; a name persists it as active
- *  and HOT-SWAPS the live provider. Returns the (possibly updated) active name +
- *  context window for the caller to thread back into the REPL state. */
-async function runModelCommand(opts: {
-  arg: string;
-  provider: OpenAICompatibleProvider;
-  activeName: string;
-  fallbackEntry: IModelEntry;
-  contextWindow: number;
-}): Promise<{ activeName: string; contextWindow: number }> {
-  const { arg, provider, activeName, fallbackEntry, contextWindow } = opts;
-  const wanted = arg.trim();
-
-  if (wanted.length === 0) {
-    await listModels(provider, activeName);
-
-    return { activeName, contextWindow };
-  }
-
-  try {
-    const next = await setActiveModel(wanted);
-    const entry = next.models[wanted] ?? fallbackEntry;
-
-    provider.reconfigure(providerConfig(entry));
-
-    const window =
-      entry.contextWindow ??
-      (await detectContextWindow(entry)) ??
-      contextWindow;
-    const info = modelInfo(provider.config);
-
-    process.stdout.write(
-      `  ✓ switched to ${wanted} — ${info.model} @ ${info.endpoint} (context ${String(window)})\n`
-    );
-
-    return { activeName: wanted, contextWindow: window };
-  } catch (err) {
-    process.stdout.write(
-      `  ${err instanceof Error ? err.message : String(err)}\n`
-    );
-
-    return { activeName, contextWindow };
-  }
-}
-
-/** List saved sessions for a directory (the `/sessions` command). */
-async function printSessions(dir: string): Promise<void> {
-  const sessions = await listSessions(dir);
-
-  if (sessions.length === 0) {
-    process.stdout.write("no saved sessions for this directory\n");
-
-    return;
-  }
-
-  for (const s of sessions) {
-    const firstUser = s.messages.find((m) => m.role === "user")?.content ?? "";
-    const snippet = firstUser.slice(0, 48).replace(/\s+/g, " ");
-
-    process.stdout.write(
-      `  ${s.id}  ${String(s.messages.length).padStart(3)} msgs  ${snippet}\n`
-    );
-  }
-}
-
 export { makeSpinner, spinnerPhase, type ISpinnerOut } from "./render/spinner";
-
-const spinner = makeSpinner();
-
-/** When the interactive REPL pins an editable input row, streamed output must be
- *  written THROUGH the StatusBar (so it scrolls in the region above the row and
- *  the cursor stays parked on the row). Null elsewhere ⇒ a plain stdout write. */
-let interactiveStream: ((text: string) => void) | null = null;
-
-const render: Reporter = (event) => {
-  const phase = spinnerPhase(event);
-
-  if (phase !== null) {
-    spinner.setLabel(phase);
-  }
-
-  const out = renderEvent(event, { color: true });
-
-  if (out.length > 0) {
-    spinner.clear();
-
-    if (interactiveStream !== null) {
-      interactiveStream(out);
-    } else {
-      process.stdout.write(out);
-    }
-  }
-};
-
-/** Reporter that renders to the terminal AND, when `--log <file>` is set, appends
- *  the full event stream as JSONL (one event per line, timestamped) for later
- *  evaluation — the durable record of what the agent did: its reasoning, every
- *  file it wrote, the gate verdicts, and the loops it got stuck in. Append-only
- *  (NOT overwritten like the session JSON), and unredacted — it's an opt-in local
- *  debug artifact. Logging failures never break the session. */
-function makeReporter(
-  logFile: string,
-  runId: string,
-  sessionId?: string
-): Reporter {
-  if (logFile.length === 0) {
-    return render;
-  }
-
-  const ledger = new LedgerWriter(logFile, runId, sessionId);
-
-  return (event) => {
-    render(event);
-
-    const { kind, ...rest } = event;
-
-    ledger.record(ledgerTypeFor(event), { kind, ...rest });
-  };
-}
-
-/** Resolve the run-log file when `--log` is set: an auto-named, timestamped JSONL
- *  under ~/.tsforge/logs/ (created if needed), so logs are always in one findable
- *  place and you never specify a path. Empty string = logging off. */
-function resolveLogPath(id: string, enabled: boolean): string {
-  if (!enabled) {
-    return "";
-  }
-
-  const dir = logsDir();
-
-  mkdirSync(dir, { recursive: true });
-
-  const stamp = new Date()
-    .toISOString()
-    .replace(/[:T]/g, "-")
-    .replace(/\..+$/, "");
-
-  return join(dir, `${stamp}-${id}.jsonl`);
-}
-
-/** The model for a run: a recipe's named model (from ~/.tsforge/models.json) when
- *  set and known, else the active model. An unknown name warns and falls back. */
-async function modelForRun(
-  args: ICliArgs
-): Promise<{ name: string; entry: IModelEntry }> {
-  if (args.model.length > 0) {
-    const cfg = await loadModelsConfig();
-    const entry = cfg.models[args.model];
-
-    if (entry !== undefined) {
-      return { name: args.model, entry };
-    }
-
-    process.stdout.write(
-      `  recipe model '${args.model}' not in models.json — using the active model\n`
-    );
-  }
-
-  return resolveActiveModel();
-}
+export { providerConfig } from "./cli/model-setup";
+export { isApproval, isPlanApproval } from "./cli/repl";
 
 /** One-shot: drive a single task to green, then exit. */
 async function runOnce(args: ICliArgs): Promise<number> {
@@ -578,1855 +112,6 @@ async function runOnce(args: ICliArgs): Promise<number> {
   }
 
   return ok ? 0 : 1;
-}
-
-/** Wide approval — the staged-web checkpoint explicitly prompted "type
- *  'approve'", so casual yeses count there. */
-export function isApproval(line: string): boolean {
-  return /^(approve|approved|ok|okay|yes|y|go|lgtm)\.?$/i.test(line.trim());
-}
-
-/** Narrow approval — GENERAL plan mode, where the model asks clarifying
- *  questions: a "yes" may ANSWER a question, so only unambiguous approval
- *  words exit the mode and start implementing. */
-export function isPlanApproval(line: string): boolean {
-  return /^(approve|approved|go|lgtm|implement)[.!]?$/i.test(line.trim());
-}
-
-// The /help body is generated from the command registry (src/cli/commands.ts) so
-// the help text and the interactive `/` palette can never drift.
-const HELP = formatHelp();
-
-/** A single compact "how to start" line under the banner — the only guidance the
- *  landing screen needs. The internals (cwd, scope, gate, session) live in /config. */
-function startupHint(): string {
-  const tip = (key: string, label: string): string =>
-    `${paint(key, STYLE.brand + STYLE.bold, true)} ${paint(label, STYLE.dim, true)}`;
-  const sep = paint("   ·   ", STYLE.dim, true);
-
-  return `  ${[
-    tip("/help", "commands"),
-    tip("@", "files"),
-    tip("/setup", "guardrails"),
-    tip("/exit", "quit"),
-  ].join(sep)}`;
-}
-
-/** The post-turn plan-mode footer — a compact styled chip (matches the startup
- *  plan line) instead of a plain full-width parenthetical. `ready` = the agent has
- *  proposed a plan (nudge toward approve); otherwise it's still exploring. */
-function planHint(ready: boolean): string {
-  const chip = paint(
-    `◆ plan${ready ? " ready" : ""}`,
-    STYLE.brand + STYLE.bold,
-    true
-  );
-  const reply = paint("reply to refine · type", STYLE.dim, true);
-  const approve = paint("approve", STYLE.green + STYLE.bold, true);
-  const tail = paint(ready ? "to build" : "when ready", STYLE.dim, true);
-
-  return `  ${chip}  ${paint("·", STYLE.dim, true)}  ${reply} ${approve} ${tail}`;
-}
-
-/** Print the welcome banner, a compact hint, and (when resuming) the prior transcript. */
-function printHeader(info: {
-  dir: string;
-  id: string;
-  gateLabel: string;
-  files: string[];
-  resumed: ISessionRecord | null;
-  model: { model: string; endpoint: string };
-  updateNotice?: string | null;
-}): void {
-  const { resumed, model, updateNotice } = info;
-
-  if (process.stdout.isTTY) {
-    // Clean slate: wipe the visible screen AND scrollback so the banner never
-    // lands on top of leftover shell output (env dumps, prior command noise).
-    process.stdout.write("\x1b[2J\x1b[3J\x1b[H");
-  }
-
-  process.stdout.write(welcomeBanner(model));
-
-  if (updateNotice !== undefined && updateNotice !== null) {
-    process.stdout.write(`${updateNotice}\n`);
-  }
-
-  process.stdout.write(`${startupHint()}\n\n`);
-
-  if (resumed === null) {
-    return;
-  }
-
-  // Replay the prior conversation so a resumed session has visible context.
-  process.stdout.write("\n── resuming conversation ──\n");
-
-  for (const message of resumed.messages) {
-    process.stdout.write(
-      renderMessage(message, { color: true, speaker: model.model })
-    );
-  }
-
-  process.stdout.write("\n──────────────────────────\n");
-}
-
-// tsforge's bundled browser-check script (headless-chromium render oracle).
-const BROWSER_CHECK = join(
-  import.meta.dir,
-  "..",
-  "scripts",
-  "browser-check.ts"
-);
-
-function browserCheckCommand(htmlFile: string): string {
-  return `bun "${BROWSER_CHECK}" "${htmlFile}"`;
-}
-
-/**
- * Resolve the session's gate + label. Starts from the base gate (resumed /
- * explicit / auto strict-TS), then appends a `--browser` render check when asked
- * — so a web build is verified to actually RUN, not just type-check.
- */
-async function resolveGate(
-  args: ICliArgs,
-  resumed: ISessionRecord | null
-): Promise<{ accept: string; gateLabel: string; lintFile?: FileLinter }> {
-  const base = await baseGate(args, resumed);
-
-  if (args.browser.length === 0) {
-    return base;
-  }
-
-  const browser = browserCheckCommand(args.browser);
-
-  return {
-    accept: base.accept.length > 0 ? `${base.accept} && ${browser}` : browser,
-    gateLabel:
-      base.accept.length > 0
-        ? `${base.gateLabel} + browser render`
-        : "browser render",
-    ...(base.lintFile === undefined ? {} : { lintFile: base.lintFile }),
-  };
-}
-
-/** The base gate: a resumed session's gate wins, then explicit `--accept`, then
- *  `--no-gate` (off), else tsforge's auto gate (strict-TS / project lint). */
-async function baseGate(
-  args: ICliArgs,
-  resumed: ISessionRecord | null
-): Promise<{ accept: string; gateLabel: string; lintFile?: FileLinter }> {
-  if (resumed !== null) {
-    const label = resumed.accept.length > 0 ? resumed.accept : "none";
-
-    return { accept: resumed.accept, gateLabel: label };
-  }
-
-  if (args.accept.length > 0) {
-    return { accept: args.accept, gateLabel: args.accept };
-  }
-
-  if (args.web) {
-    // The --web SCAFFOLD path is greenfield: tsforge writes the skeleton in its
-    // own house style, so the web gate + web guidance deliberately stay on the
-    // defaults and do NOT thread project `conventions` (which govern the core
-    // brownfield path). Keeping both on house style avoids a gate/guidance
-    // contradiction. See docs/harness-subsystems.md "setup / conventions".
-    const web = buildWebGate("react", undefined, args.dir);
-
-    // PER-WRITE lint moat: the web gate's eslint rules applied to each file as the
-    // model writes it, so architecture/cast violations surface immediately instead
-    // of as an end-of-turn pile-up.
-    return {
-      accept: web.command,
-      gateLabel: web.label,
-      lintFile: makeFileLinter("react", args.dir, WEB_PACKS),
-    };
-  }
-
-  if (args.noGate) {
-    return { accept: "", gateLabel: "none (--no-gate)" };
-  }
-
-  const { detectStack } = await import("./stack-detection");
-  const {
-    loadTsforgeConfig,
-    resolveActivePacks,
-    normalizeRuleOverrides,
-    resolveProjectProfile,
-  } = await import("./config/tsforge-config");
-  const { resolveConventions } = await import("./infer-rules/conventions");
-
-  const stackProfile = await detectStack(args.dir);
-  const config = await loadTsforgeConfig(args.dir);
-  const activePacks = resolveActivePacks(stackProfile.packs, config);
-  const ruleOverrides = normalizeRuleOverrides(config);
-  const profile = resolveProjectProfile(config);
-  const conventions = resolveConventions(config.conventions);
-
-  const auto = await buildGate(
-    args.dir,
-    activePacks,
-    Object.keys(ruleOverrides).length > 0 ? ruleOverrides : undefined,
-    {
-      enableTypeAware: profile === "strict",
-      // "Green" should mean the strict floor AND the project's own tests pass —
-      // not just that it type-checks and lints. discoverTestCommand appends them
-      // only when the project actually has tests; --strict-floor-only opts out.
-      includeTests: !args.strictFloorOnly,
-      conventions,
-    }
-  );
-
-  return {
-    accept: auto.command,
-    gateLabel: auto.label,
-    lintFile: makeFileLinter(
-      "core",
-      args.dir,
-      activePacks,
-      Object.keys(ruleOverrides).length > 0 ? ruleOverrides : undefined,
-      conventions
-    ),
-  };
-}
-
-/** One-line nudge when the repo has no config yet — setup adapts the guardrails
- *  to this repo's conventions. Just a hint; never auto-runs. */
-function maybePrintNoConfigHint(
-  dir: string,
-  resumed: ISessionRecord | null
-): void {
-  if (resumed === null && !existsSync(join(dir, "tsforge.config.json"))) {
-    const icon = paint("○", STYLE.yellow, true);
-    const run = paint("/setup", STYLE.brand + STYLE.bold, true);
-    const rest = paint("to adapt the guardrails to this repo", STYLE.dim, true);
-
-    process.stdout.write(`  ${icon} no project config — run ${run} ${rest}\n`);
-  }
-}
-
-/** Initialize the REPL session: resolve model, gate, context window, and create
- *  the session object. Returns the session, provider, and config metadata.
- *  Extracted to reduce repl() cognitive complexity. */
-async function initReplSession(args: ICliArgs): Promise<{
-  session: Session;
-  provider: OpenAICompatibleProvider;
-  activeName: string;
-  contextWindow: number;
-  id: string;
-  gateLabel: string;
-  logFile: string;
-  persist: () => Promise<void>;
-  report: Reporter;
-  resumed: ISessionRecord | null;
-  files: string[];
-  activeModelEntry: IModelEntry;
-}> {
-  const activeModel = await modelForRun(args);
-  const provider = makeProvider(activeModel.entry);
-  const activeName = activeModel.name;
-
-  warnDefaultModelOnRemote(activeModel.entry);
-
-  // Best-effort cleanup of stale sessions on every launch.
-  await pruneSessions();
-
-  // --resume <id> loads a specific session; --continue the newest for this dir.
-  const resumed =
-    args.resumeId.length > 0
-      ? await loadSession(args.resumeId)
-      : args.continue
-        ? await latestSession(args.dir)
-        : null;
-
-  if ((args.continue || args.resumeId.length > 0) && resumed === null) {
-    process.stdout.write("(no matching saved session — starting fresh)\n");
-  }
-
-  // --web: lay down the opinionated skeleton before resolving the gate.
-  if (args.web && resumed === null) {
-    await setUpWebProject(args.dir, "react");
-  }
-
-  const id = resumed?.id ?? newSessionId();
-  const { accept, gateLabel, lintFile } = await resolveGate(args, resumed);
-  const files = resumed !== null ? resumed.files : scopeOf(args);
-  const logFile = resolveLogPath(id, args.log);
-
-  if (logFile.length > 0) {
-    process.stdout.write(`  ↳ logging this run to ${logFile}\n`);
-  }
-
-  // Scout seeds a one-shot drive-to-green run's first prompt; interactive sessions
-  // gather context conversationally, so it doesn't apply here. Say so rather than
-  // silently ignore the flag.
-  if (args.scout) {
-    process.stdout.write(
-      '  ↳ note: --scout applies to one-shot runs (tsforge "task" --files … --scout); ignored in interactive mode\n'
-    );
-  }
-
-  const thinkingTokenBudget = envNumber("TSFORGE_THINKING_BUDGET");
-  // Auto-compaction threshold (fraction of the window); session default 0.8.
-  const autoCompactAt = envNumber("TSFORGE_COMPACT_AT");
-  // The model's real context window: explicit env wins, else ask the server
-  // (max_model_len), else a conservative fallback. Drives the status gauge AND
-  // auto-compaction (the session compacts before a send once it nears the window).
-  const contextWindow =
-    activeModel.entry.contextWindow ??
-    envNumber("TSFORGE_CONTEXT_WINDOW") ??
-    (await detectContextWindow(provider.config)) ??
-    32_768;
-  const report = makeReporter(logFile, id, id);
-  const config = {
-    provider,
-    cwd: args.dir,
-    files,
-    accept,
-    contextWindow,
-    report,
-    // PER-WRITE lint moat (eslint rules per file as it's written), so violations
-    // surface immediately instead of piling up at the end-of-turn gate.
-    ...(lintFile === undefined ? {} : { lintFile }),
-    ...(resumed === null ? {} : { history: resumed.messages }),
-    // --web pre-scaffolds the project above, so it gets the web gate/guidance
-    // directly. EVERY OTHER interactive session offers `scaffold_web` (+ the
-    // ui/routes tools that ride along) so the AGENT can decide mid-conversation
-    // that a request is a from-scratch web app — this flag is what puts the tool
-    // in the model's list; setSetupWeb() below only wires its callback.
-    ...(args.web
-      ? {
-          // --web pre-scaffolds the app, so scaffold_web isn't needed — but the
-          // build still needs scaffold_ui + scaffold_routes (+ add_dependency),
-          // which `scaffoldUi: true` registers. Without this the web guidance
-          // tells the model to call tools that aren't in its list and it deadlocks.
-          scaffoldUi: true,
-          guidance: webGuidance("react"),
-          fix: buildWebFix("react"),
-          incrementalCheck: buildWebTscCheck(args.dir),
-        }
-      : { scaffoldWeb: true, fix: buildCoreFix() }),
-    ...(thinkingTokenBudget === undefined ? {} : { thinkingTokenBudget }),
-    ...(autoCompactAt === undefined ? {} : { autoCompactAt }),
-    // `--policy-mode` (validated) overrides the config file's policy.mode.
-    ...(isPolicyMode(args.policyMode) ? { policyMode: args.policyMode } : {}),
-    // Thinking OFF for interactive replies so they STREAM immediately instead of
-    // stalling on a long hidden chain-of-thought (qwen-local defaults thinking on).
-    // The session still flips thinking ON automatically while repairing gate errors.
-    enableThinking: false,
-  };
-
-  const session = await Session.create(config);
-
-  // A self-describing run-meta line at the top of the --log so the analyzer knows
-  // which model / context window the metrics are against (the thread's advice:
-  // many "model failures" are really quant/config failures — record the config).
-  report({
-    kind: "start",
-    task: "session",
-    message: `model ${modelInfo(provider.config).model} · context window ${contextWindow}`,
-    model: modelInfo(provider.config).model,
-    contextWindow,
-  });
-
-  const persist = async (): Promise<void> => {
-    await saveSession({
-      id,
-      cwd: args.dir,
-      // The LIVE gate/scope — not the startup constants. /gate, /files, and a web
-      // scaffold all mutate these mid-session; persisting the originals would
-      // silently restore stale settings on --continue. See P2 review.
-      accept: session.gate,
-      files: session.scope,
-      updatedAt: Date.now(),
-      planMode: false, // will be set by caller
-      messages: [...session.messages],
-    });
-  };
-
-  return {
-    session,
-    provider,
-    activeName,
-    contextWindow,
-    id,
-    gateLabel,
-    logFile,
-    persist,
-    report,
-    resumed,
-    files,
-    activeModelEntry: activeModel.entry,
-  };
-}
-
-/** Interactive REPL: a persistent gate-anchored conversation. */
-async function repl(args: ICliArgs): Promise<number> {
-  // Interactive sessions get web tools ON by default (an assistant that can't look
-  // things up is silly). Only a DEFAULT — an explicit TSFORGE_WEB (incl. "0") wins,
-  // and one-shot/headless/eval never run this path, so they stay offline+deterministic.
-  process.env.TSFORGE_WEB ??= "1";
-
-  const {
-    session: initialSession,
-    provider,
-    activeName: initialActiveName,
-    contextWindow: initialContextWindow,
-    id,
-    gateLabel: initialGateLabel,
-    logFile,
-    resumed,
-    files,
-    activeModelEntry,
-  } = await initReplSession(args);
-
-  let session = initialSession;
-  let activeName = initialActiveName;
-  let contextWindow = initialContextWindow;
-  // A human label for the gate (e.g. "strict TypeScript / project lint"), shown in
-  // the header + /config instead of the raw multi-line command. Updated when the
-  // user sets a gate via /config.
-  let gateLabel = initialGateLabel;
-
-  const persist = async (): Promise<void> => {
-    await saveSession({
-      id,
-      cwd: args.dir,
-      // The LIVE gate/scope — not the startup constants. /gate, /files, and a web
-      // scaffold all mutate these mid-session; persisting the originals would
-      // silently restore stale settings on --continue. See P2 review.
-      accept: session.gate,
-      files: session.scope,
-      updatedAt: Date.now(),
-      planMode,
-      messages: [...session.messages],
-    });
-  };
-
-  // "update available" notice: read from the local cache (no network on the hot
-  // path) and refresh it in the background for next time. Gated to interactive,
-  // non-CI sessions inside update-check, so eval/headless runs are unaffected.
-  const updateNotice = await getUpdateNotice(currentVersion());
-
-  refreshUpdateCacheInBackground();
-
-  printHeader({
-    dir: args.dir,
-    id,
-    gateLabel,
-    files,
-    resumed,
-    model: modelInfo(provider.config),
-    updateNotice,
-  });
-
-  maybePrintNoConfigHint(args.dir, resumed);
-
-  // Pin an editable input row only on a real TTY tall enough to host the bar.
-  // In that mode readline does line-EDITING but must not RENDER (we paint the
-  // row ourselves), so it gets a discard sink for output; otherwise it writes to
-  // stdout as before (pipes, small terminals — behaviour unchanged).
-  const useInputRow =
-    process.stdin.isTTY &&
-    process.stdout.isTTY &&
-    process.stdout.rows >= MIN_ROWS;
-
-  // In editor mode, do NOT create readline — the editor owns stdin exclusively.
-  // In fallback mode (non-TTY or basicInput), readline is the only consumer.
-  const useEditor = useInputRow && !flags.basicInput();
-
-  const inputSink = new Writable({
-    write(_chunk, _enc, cb): void {
-      cb();
-    },
-  });
-
-  const rl = useEditor
-    ? null
-    : createInterface({
-        input: process.stdin,
-        output: useInputRow ? inputSink : process.stdout,
-        terminal: true,
-      });
-
-  // Ctrl-C: while a turn is running, abort it and return to the prompt; while
-  // idle at the prompt, quit. (readline emits SIGINT on the interface, so the
-  // process isn't killed — we decide what it means.)
-  let active: AbortController | null = null;
-  // Lines typed WHILE a run is in flight — drained at each turn boundary to steer
-  // the model (see Session.send `steer`), instead of blocking until the run ends.
-  const pending: string[] = [];
-
-  if (rl !== null) {
-    rl.on("SIGINT", () => {
-      if (active !== null) {
-        active.abort();
-      } else {
-        rl.close();
-      }
-    });
-  }
-
-  // Explicit `--web` (no Q&A): the FIRST message is the build, so stage it
-  // (plan+types → implement). Cleared after, so follow-ups are plain sends.
-  let stagedWebPending = args.web && resumed === null;
-  // Plan mode is the DEFAULT for a fresh interactive session (opt out with
-  // `--no-plan` or an explicit non-plan `--policy-mode`/config `policy.mode`).
-  // For a staged web build it pauses after the design phase to review the plan;
-  // for EVERYTHING else it is the general read-only mode: the agent explores,
-  // asks clarifying questions, and proposes a plan — only an explicit approval
-  // unlocks tools and implements. A resumed session restores its saved mode
-  // (the read-only guarantee must survive `--continue`).
-  let planMode = resolveInitialPlanMode(
-    args,
-    resumed?.planMode,
-    session.basePolicyMode
-  );
-  // True once a plan-mode exchange has happened, so a stray "approve" before any
-  // discussion is just a message, not an approval.
-  let planDiscussed = false;
-  // The current interactive mode (Shift+Tab cycles it; /plan toggles it). Kept in
-  // sync with `planMode`; shown as a chip in the status bar.
-  let currentModeId = planMode ? "plan" : "normal";
-
-  session.setPlanMode(planMode);
-
-  if (planMode) {
-    const chip = paint("◆ plan mode (default)", STYLE.brand + STYLE.bold, true);
-    const body = paint(
-      "— I'll explore and propose a plan; reply",
-      STYLE.dim,
-      true
-    );
-    const approve = paint("approve", STYLE.green + STYLE.bold, true);
-    const tail = paint("to build", STYLE.dim, true);
-
-    process.stdout.write(`  ${chip} ${body} ${approve} ${tail}\n`);
-  }
-
-  // While set, the next user line is the plan-review reply ("approve", or edits to
-  // fold into phase 2) — the design phase has run and is waiting at the checkpoint.
-  let awaitingPlanApproval = false;
-
-  const configureWeb = async (
-    framework: WebFramework,
-    options: { signal?: AbortSignal } = {}
-  ): Promise<{ files: readonly string[]; depsInstalled: boolean }> => {
-    process.stdout.write(
-      `\n  ↳ scaffolding a ${frameworkLabel(framework)} project\n`
-    );
-
-    const setup = await setUpWebProject(args.dir, framework, options);
-
-    session.setGate(buildWebGate(framework, undefined, args.dir).command);
-    session.setFix(buildWebFix(framework));
-    session.setIncrementalCheck(buildWebTscCheck(args.dir));
-    // The project only now has a tsconfig + node_modules — rebuild the TS service
-    // so the per-write guard actually runs (it's skipped on a null service), and
-    // switch the lint moat to the web rules so component-architecture /
-    // no-jsx-computation / cast violations surface per file, not at the gate.
-    await session.refreshTsService();
-    session.setLintFile(makeFileLinter(framework, args.dir, WEB_PACKS));
-    session.guide(webGuidance(framework));
-    // A from-scratch web build legitimately needs many turns. Don't pin a low
-    // ceiling here — the interactive session already rides the high runaway
-    // backstop (interactiveBackstopTurns) and stops on the progress guards, so a
-    // long, converging build is never cut off mid-write.
-
-    return setup;
-  };
-
-  // The `scaffold_web` tool invokes this when the AGENT decides to build a web app
-  // (the framework string is validated tool-side). `configureWeb` closes over the
-  // mutable `session`, so this stays correct across `/clear`; re-applied below.
-  const setupWeb: SetupWebFn = (framework, options) =>
-    configureWeb(framework === "vanilla" ? "vanilla" : "react", options);
-
-  session.setSetupWeb(setupWeb);
-
-  // Last-turn summary, surfaced in the status line shown before each prompt.
-  let lastTurns = 0;
-  // Turns the last GREEN run took (the loop-efficiency signal shown in /metrics).
-  let lastTurnsToGreen: number | null = null;
-  let lastElapsedMs = 0;
-  let lastStatus = "ready";
-
-  // Run one user-driven exchange: fresh abort controller, time it, record the
-  // outcome for the status line, persist. `run` gets the live signal + a steer
-  // drain so in-flight user messages reach the model.
-  const drive = async (
-    run: (opts: { signal: AbortSignal; steer: () => string[] }) => Promise<{
-      status: string;
-      turns: number;
-    }>
-  ): Promise<void> => {
-    active = new AbortController();
-    const started = performance.now();
-
-    lastStatus = "working"; // reflected live on the bar (● working) during the turn
-    spinner.start();
-
-    try {
-      const result = await run({
-        signal: active.signal,
-        steer: () => pending.splice(0, pending.length),
-      });
-
-      lastTurns = result.turns;
-
-      if (result.status === "done") {
-        lastTurnsToGreen = result.turns;
-      }
-
-      lastElapsedMs = performance.now() - started;
-      lastStatus = result.status;
-    } finally {
-      spinner.stop();
-      active = null;
-      // Seal the agent card's `╰` bottom cap the moment streaming ends, so any
-      // post-turn hint (plan-mode notice, PLAN review, etc.) lands BELOW the card
-      // instead of inside it — which would break the rail. Idempotent.
-      closeAgentTurn();
-    }
-
-    await persist();
-  };
-
-  // Free-text user sends route through here: resolve `@file` mentions to inlined
-  // contents (composeMessage) before handing the message to the session. The
-  // plan-approval / staged-build sends call session.send directly and are not
-  // touched, so only ordinary messages get mention expansion.
-  const runSend = (line: string): Promise<void> =>
-    drive(async (opts) =>
-      session.send(await composeMessage(args.dir, line), opts)
-    );
-
-  // A from-scratch web build: stage it (plan + types, then implement) so the
-  // model designs the type contract before writing UI — far less API invention.
-  // The design phase gates on TYPES only (tsc + lint) so contract errors surface
-  // early and small, not as a final avalanche. `withPlan` is the web flow's OWN
-  // checkpoint (design writes types, so general read-only plan mode must be off).
-  const runStagedBuild = (
-    line: string,
-    framework: WebFramework,
-    withPlan: boolean
-  ): Promise<void> =>
-    withPlan
-      ? runPlanned(line, framework)
-      : drive((opts) =>
-          session.buildStaged(
-            line,
-            opts,
-            buildWebTypeGate(framework, undefined, args.dir).command
-          )
-        );
-
-  // Plan mode: run the design phase, then show the model's plan and PAUSE — the
-  // next user line approves it (or edits it, folded into phase 2). The design runs
-  // inside drive() (signal/steer/persist); the quick plan summary is captured for
-  // the prompt that follows.
-  const runPlanned = async (
-    line: string,
-    framework: WebFramework
-  ): Promise<void> => {
-    let plan = "";
-
-    await drive(async (opts) => {
-      const designed = await session.designBuild(
-        line,
-        opts,
-        buildWebTypeGate(framework, undefined, args.dir).command
-      );
-
-      if (designed.status !== "interrupted") {
-        plan = await session.generatePlan();
-      }
-
-      return designed;
-    });
-
-    if (plan.length > 0) {
-      echo(
-        `\n📋 PLAN — review, then type 'approve' to build, or describe changes:\n\n${plan}\n\n`
-      );
-      awaitingPlanApproval = true;
-    }
-  };
-
-  const dispatch = async (line: string): Promise<void> => {
-    // A reply to the plan checkpoint: "approve" (build as-planned) or any other
-    // text = corrections folded into the implement phase. Either way phase 2 runs.
-    if (awaitingPlanApproval) {
-      awaitingPlanApproval = false;
-
-      const approved = isApproval(line);
-      const notes = approved ? "" : line;
-
-      if (!approved) {
-        echo("  ↳ folding your changes into the build\n");
-      }
-
-      await drive((opts) => session.implementBuild(notes, opts));
-
-      return;
-    }
-
-    // Explicit --web: the first message is a from-scratch build — stage it. The
-    // staged flow has its OWN plan checkpoint (its design phase writes types),
-    // so general read-only plan mode hands over to it here.
-    if (stagedWebPending) {
-      stagedWebPending = false;
-
-      const withPlan = planMode;
-
-      planMode = false;
-      planDiscussed = false;
-      session.setPlanMode(false);
-      await runStagedBuild(line, "react", withPlan);
-
-      return;
-    }
-
-    // GENERAL plan mode, approval: unlock the tools and implement the plan that
-    // is already the latest assistant message. Only an explicit approval word
-    // counts ("yes" may be answering one of the model's clarifying questions).
-    if (planMode && planDiscussed && isPlanApproval(line)) {
-      planMode = false;
-      planDiscussed = false;
-      session.setPlanMode(false);
-      echo("  ✓ plan approved — implementing\n");
-      await drive((opts) => session.send(PLAN_APPROVED_NOTE, opts));
-
-      return;
-    }
-
-    // GENERAL plan mode, discussion: the agent explores read-only, asks its
-    // clarifying questions, and proposes/revises a plan. Stays in plan mode.
-    if (planMode) {
-      await runSend(line);
-      planDiscussed = true;
-
-      const last = session.messages.at(-1);
-      const planned =
-        last?.role === "assistant" && /^##\s*plan\b/im.test(last.content);
-
-      echo(`\n${planHint(planned)}\n`);
-
-      return;
-    }
-
-    // No up-front classifier: the AGENT decides. It calls `scaffold_web` itself
-    // when the request is a from-scratch web app, and just answers/edits otherwise
-    // (so "render a table in the CLI" is no longer mis-scaffolded as a Vite app).
-    await runSend(line);
-  };
-
-  // Placeholder declaration for handleHelp; defined after runLine is available.
-  let handleHelp: () => Promise<void>;
-
-  // Slash-command dispatch. Returns true to EXIT the REPL. Kept as a closure so
-  // it can rebuild `session` (e.g. /clear) and reach config/persist.
-  const command = async (line: string): Promise<boolean> => {
-    const [verb, ...rest] = line.slice(1).split(" ");
-    const arg = rest.join(" ").trim();
-
-    switch ((verb ?? "").toLowerCase()) {
-      case "exit":
-      case "quit":
-        return true;
-      case "help":
-        await handleHelp();
-        break;
-      case "clear":
-        // Rebuild the session with the current state (config is not reused;
-        // repl's /clear creates a fresh Session.create call)
-        session = await Session.create({
-          provider,
-          cwd: args.dir,
-          files: session.scope,
-          accept: session.gate,
-          contextWindow,
-          report: makeReporter(logFile, id, id),
-          enableThinking: false,
-        });
-        session.setSetupWeb(setupWeb);
-        session.setPlanMode(planMode); // a /clear must not silently drop the mode
-        planDiscussed = false;
-        await persist();
-        clearScreen(); // wipe the visible terminal + scrollback, not just the state
-        process.stdout.write("conversation cleared\n");
-        break;
-
-      case "compact": {
-        // Compaction is a full model round-trip (can take many seconds). Drive the
-        // SAME live-activity path a turn uses: lastStatus → "● working" on the bar,
-        // spinner.start() runs the tick timer whose onTick repaints the bar with the
-        // "⠋ compacting · Ns" activity segment (the inline spinner is suppressed in
-        // the REPL, so the bar IS the loader). ALWAYS restore + stop, even on a
-        // provider error, so the prompt comes back clean and idle.
-        lastStatus = "working";
-        spinner.start();
-        spinner.setLabel("compacting");
-
-        try {
-          const { before, after } = await session.compact();
-
-          await persist();
-          process.stdout.write(`compacted ${before} → ${after} messages\n`);
-        } finally {
-          spinner.stop();
-          lastStatus = "ready";
-        }
-
-        break;
-      }
-
-      case "plan":
-        togglePlanMode();
-        break;
-
-      case "gate":
-        session.setGate(arg);
-        process.stdout.write(
-          arg.length > 0 ? `gate: ${arg}\n` : "gate cleared\n"
-        );
-        // Persist immediately so a `/gate` change survives even if the user quits
-        // before the next send (persist otherwise only runs after a turn).
-        await persist();
-        break;
-
-      case "review":
-        await runReviewCommand(provider, args.dir, arg);
-        break;
-
-      case "map":
-        await runMapCommand(args.dir, arg);
-        break;
-
-      case "trace":
-        await runTraceCommand(arg, logFile);
-        break;
-
-      case "config":
-        await handleConfig();
-        break;
-
-      case "setup": {
-        const { runSetup } = await import("./setup/run-setup");
-
-        // runSetup prints its own apply/cancel summary — don't add a second,
-        // possibly-misleading line (it would claim success even on cancel).
-        await runSetup({
-          cwd: args.dir,
-          yes: false,
-          color: process.stdout.isTTY,
-          // The REPL editor/readline owns stdin — don't let the wizard pause it
-          // on exit (that would quit the whole process).
-          manageInput: false,
-        });
-        break;
-      }
-
-      case "files": {
-        const globs = arg
-          .split(",")
-          .map((s) => s.trim())
-          .filter(Boolean);
-
-        session.setScope(globs.length > 0 ? globs : WHOLE_REPO);
-        process.stdout.write(`scope: ${scopeLabel(session.scope)}\n`);
-        await persist();
-        break;
-      }
-
-      case "model": {
-        const result = await runModelCommand({
-          arg,
-          provider,
-          activeName,
-          fallbackEntry: activeModelEntry,
-          contextWindow,
-        });
-
-        activeName = result.activeName;
-        contextWindow = result.contextWindow;
-        // Keep auto-compaction in sync with the new model's window — not just the
-        // status bar. Otherwise a swap to a smaller model compacts too late.
-        session.setContextWindow(contextWindow);
-        break;
-      }
-
-      case "sessions":
-        await printSessions(args.dir);
-        break;
-
-      case "memory": {
-        if (arg.trim() === "forget") {
-          await forgetMemory(args.dir);
-          process.stdout.write("  memory cleared for this repo\n");
-          break;
-        }
-
-        const ledger = await loadLedger(args.dir);
-
-        if (ledger.entries.length === 0) {
-          process.stdout.write("  no learned lessons yet\n");
-          break;
-        }
-
-        const activeNames = new Set(
-          activeRules(ledger, Date.now()).map((r) => r.name)
-        );
-
-        process.stdout.write(
-          `  ${String(ledger.entries.length)} lesson(s), ${String(activeNames.size)} active (● fires · ○ still accruing):\n`
-        );
-
-        for (const entry of ledger.entries.slice(0, 20)) {
-          const mark = activeNames.has(entry.name) ? "●" : "○";
-
-          process.stdout.write(
-            `    ${mark} ${entry.rule} · ${String(entry.hits)} hit(s)\n`
-          );
-        }
-
-        process.stdout.write("  /memory forget to clear\n");
-        break;
-      }
-
-      case "cost": {
-        const chars = session.messages.reduce(
-          (sum, m) => sum + m.content.length,
-          0
-        );
-
-        process.stdout.write(
-          `  ${String(session.messages.length)} messages · ~${String(Math.round(chars / 4))} tokens (rough)\n`
-        );
-        break;
-      }
-
-      case "metrics": {
-        const m = session.metrics;
-
-        if (m.calls === 0) {
-          process.stdout.write("  no model calls yet\n");
-        } else {
-          process.stdout.write(
-            `  ${String(m.calls)} call(s) · ${String(m.promptTokens)} in / ${String(m.completionTokens)} out · ` +
-              `${String(m.lastTokensPerSecond)} tok/s last · ${String(m.avgTokensPerSecond)} tok/s avg\n`
-          );
-        }
-
-        process.stdout.write(turnsToGreenLine(lastTurnsToGreen));
-
-        break;
-      }
-
-      default:
-        process.stdout.write(`unknown command: ${line} (try /help)\n`);
-    }
-
-    return false;
-  };
-
-  // Current state as the status surface sees it — shared by the pinned bar and
-  // the inline fallback so both show identical content.
-  const statusInfo = (): IStatusInfo => ({
-    model: modelInfo(provider.config).model,
-    contextTokens: session.contextTokens,
-    contextWindow,
-    turns: lastTurns,
-    elapsedMs: lastElapsedMs,
-    status: lastStatus,
-    scope: scopeLabel(session.scope),
-    mode: modeById(currentModeId).label,
-    tokensPerSecond: session.metrics.lastTokensPerSecond,
-    ...(spinner.frameLabel().length > 0
-      ? { activity: spinner.frameLabel() }
-      : {}),
-  });
-
-  // Pinned bottom status bar when we're on a real terminal; otherwise the bar is
-  // inactive and `prompt()` falls back to the inline status line (pipes, --log).
-  const statusBar = new StatusBar(process.stdout, true, true, useInputRow);
-
-  // Switch the interactive mode (via the extensible registry) and reflect it in
-  // the status bar. The single entry point for /plan, Shift+Tab, and startup —
-  // so `planMode`, `currentModeId`, and the bar never drift apart.
-  const setMode = (id: string): void => {
-    const mode = modeById(id);
-
-    mode.apply(session);
-    currentModeId = mode.id;
-    planMode = mode.id === "plan";
-    planDiscussed = false;
-
-    if (statusBar.active) {
-      statusBar.update(statusInfo());
-    }
-  };
-
-  // `/plan` toggles between plan and normal. Extracted so the slash-command
-  // dispatcher stays under the cognitive-complexity cap.
-  const togglePlanMode = (): void => {
-    const turningOn = !planMode;
-
-    setMode(turningOn ? "plan" : "normal");
-    process.stdout.write(
-      turningOn
-        ? "plan mode ON — read-only: the agent explores, asks, and proposes " +
-            "a plan; type 'approve' to implement\n"
-        : "plan mode OFF\n"
-    );
-  };
-
-  // `/config` — the in-harness settings hub. Runs as one owned-stdin menu loop;
-  // extracted from the dispatcher to keep it under the complexity cap.
-  const setEnv = (name: string, value: string | undefined): void => {
-    if (value === undefined) {
-      Reflect.deleteProperty(process.env, name);
-    } else {
-      process.env[name] = value;
-    }
-  };
-
-  const handleConfig = async (): Promise<void> => {
-    editorControl?.suspend();
-    editorControl?.setInputInert(true);
-
-    try {
-      await runConfigMenu({
-        color: process.stdout.isTTY,
-        suspend: () => {
-          editorControl?.suspend();
-          editorControl?.setInputInert(true);
-        },
-        resume: () => {
-          editorControl?.setInputInert(false);
-          editorControl?.resume();
-          editorControl?.getBuffer().setText("");
-        },
-        reconfigure: (entry) => {
-          provider.reconfigure(providerConfig(entry));
-        },
-        currentModelName: () => activeName,
-        onModelChange: (name) => {
-          activeName = name;
-        },
-        currentMode: () => modeById(currentModeId).label,
-        setMode,
-        getGate: () => gateLabel,
-        setGate: (cmd) => {
-          const trimmed = cmd.trim();
-
-          session.setGate(trimmed);
-          gateLabel = trimmed.length === 0 ? "none" : trimmed;
-        },
-        getScope: () => scopeLabel(session.scope),
-        setScope: (globs) => {
-          const parts = globs
-            .split(",")
-            .map((s) => s.trim())
-            .filter(Boolean);
-
-          session.setScope(parts.length > 0 ? parts : WHOLE_REPO);
-        },
-        getEnv: (name) => process.env[name],
-        setEnv,
-        view: {
-          render: (lines) => {
-            statusBar.setOverlay(lines, statusInfo());
-          },
-          close: () => {
-            statusBar.clearOverlay(statusInfo());
-          },
-        },
-      });
-    } finally {
-      editorControl?.setInputInert(false);
-      editorControl?.resume();
-      editorControl?.getBuffer().setText("");
-    }
-
-    if (statusBar.active) {
-      statusBar.update(statusInfo());
-    }
-
-    await persist();
-  };
-
-  // Set once the multi-line editor is created (it lives in a nested scope); the
-  // resize handler below calls it so the editor re-wraps/re-windows at the new
-  // size instead of clipping the current line at its pre-resize dimensions.
-  let resizeEditor: ((columns: number, rows: number) => void) | null = null;
-  // The live editor handle, exposed to repl-scope closures (e.g. the `/config`
-  // command) so they can suspend/resume its stdin ownership around an overlay
-  // wizard — the editor itself is created inside the loop's nested scope.
-  let editorControl: IEditorHandle | null = null;
-
-  // Each agent turn renders as a left-accent card: a rounded `╭ <model>` cap, every
-  // body line prefixed with the `│ ` rail (wrapping inside it), and a `╰` cap when
-  // the turn ends. The cap is emitted once, on the turn's first streamed output.
-  // The card's content budget leaves the rail (2) + 2 spare columns, so no terminal
-  // — however it treats the right margin — ever wraps a row and drops the rail.
-  const railInnerWidth = (): number =>
-    (process.stdout.columns > 0 ? process.stdout.columns : 80) -
-    PROMPT_COLS -
-    2;
-  let agentTurnOpen = false;
-  let agentRail = makeAgentRail(agentBar(true), railInnerWidth);
-
-  // Route streamed agent output through the bar so it scrolls above the pinned
-  // input row; cleared on loop exit so later/headless writes go straight to stdout.
-  if (useInputRow) {
-    interactiveStream = (text): void => {
-      if (!agentTurnOpen) {
-        agentTurnOpen = true;
-        agentRail = makeAgentRail(agentBar(true), railInnerWidth); // fresh per turn
-        statusBar.writeStream(`\n${agentCardTop(statusInfo().model, true)}\n`);
-      }
-
-      statusBar.writeStream(agentRail.feed(text));
-    };
-  }
-
-  // Start a fresh agent card for each turn (the cap re-emits on its first output).
-  const beginAgentTurn = (): void => {
-    agentTurnOpen = false;
-  };
-
-  // Close the current agent card (rounded bottom cap) once its turn is done. A
-  // no-op for turns that produced no streamed output (e.g. slash commands).
-  const closeAgentTurn = (): void => {
-    if (agentTurnOpen && useInputRow) {
-      statusBar.writeStream(`${agentCardBottom(true)}\n`);
-      agentTurnOpen = false;
-    }
-  };
-
-  // Mirror readline's buffer onto the input row after each keypress. setImmediate
-  // lets readline update rl.line/rl.cursor first (it processes the key async).
-  const syncInput = (): void => {
-    if (useInputRow && rl !== null) {
-      setImmediate(() => {
-        statusBar.setInput(rl.line, rl.cursor);
-      });
-    }
-  };
-
-  // Echo a CLI-side line (queued-steer notice, etc.) into the scroll region so it
-  // doesn't clobber the pinned input row; plain write when the row isn't active.
-  const echo = (text: string): void => {
-    if (useInputRow) {
-      statusBar.writeStream(text);
-    } else {
-      process.stdout.write(text);
-    }
-  };
-
-  // In the interactive REPL a readline prompt owns stdin for the WHOLE session, so
-  // the spinner's carriage-return inline write would clobber whatever the user is
-  // typing mid-turn — regardless of whether the pinned bar is active. So suppress
-  // the inline write unconditionally here: when the bar is up (≥5 rows) it shows the
-  // activity itself via statusInfo; on a sub-5-row TTY there's simply no inline
-  // spinner (correct — better silent than corrupting the input line). The default
-  // `() => true` gate still applies to any non-interactive spinner use.
-  spinner.setInlineGate(() => false);
-
-  // A drag-resize fires SIGWINCH continuously while the terminal reflows. Painting
-  // the bar into that moving target strands copies of it (the multi-bar / stray-rule
-  // mess a circular corner-drag produced). So we DEBOUNCE: while resizes are still
-  // arriving we suppress ALL bar repaints (spinner ticks included) and repaint once,
-  // cleanly, only after the size settles (~120ms of quiet).
-  const RESIZE_SETTLE_MS = 120;
-  let resizing = false;
-  let resizeTimer: ReturnType<typeof setTimeout> | null = null;
-
-  // Repaint the bar on every spinner tick so tok/s and the context meter update
-  // live mid-turn (both read live session state) — but NOT during a resize storm.
-  spinner.onTick(() => {
-    if (statusBar.active && !resizing) {
-      statusBar.update(statusInfo());
-    }
-  });
-
-  // Named so it can be detached on loop exit (an anonymous listener on the
-  // global process.stdout would pin the whole REPL closure for the process
-  // lifetime). columns/rows are typed `number` here, so no nullish guard is
-  // needed; the editor's resize ignores non-positive values regardless.
-  const handleResize = (): void => {
-    resizing = true;
-    statusBar.pauseForResize(); // buffer streamed output; draw nothing mid-storm
-
-    if (resizeTimer !== null) {
-      clearTimeout(resizeTimer);
-    }
-
-    resizeTimer = setTimeout(() => {
-      resizing = false;
-      resizeTimer = null;
-      statusBar.resize(statusInfo());
-      // The editor wraps/windows at the dimensions it was created with; without
-      // this it keeps using the pre-resize size and can clip the current line.
-      resizeEditor?.(process.stdout.columns, process.stdout.rows);
-      statusBar.flushStream(); // replay buffered output into the settled region
-    }, RESIZE_SETTLE_MS);
-  };
-
-  process.stdout.on("resize", handleResize);
-
-  // Restore the terminal even on an unexpected exit (teardown is idempotent).
-  process.on("exit", () => {
-    statusBar.teardown();
-  });
-
-  // Wipe the visible terminal + scrollback (2J + 3J + home), re-pinning the status
-  // bar around it so its scroll region stays correct. Used by /clear so the screen
-  // is a clean slate, not just the conversation state.
-  const clearScreen = (): void => {
-    const wasActive = statusBar.active;
-
-    if (wasActive) {
-      statusBar.teardown();
-    }
-
-    process.stdout.write("\x1b[2J\x1b[3J\x1b[H");
-
-    if (wasActive) {
-      statusBar.install(statusInfo());
-    }
-  };
-
-  // The prompt. With the editable input row pinned it's always visible, so we
-  // just repaint the bar + row; with the bar (no input row) it shows the inline
-  // marker; otherwise it prints the inline status line above the marker.
-  const prompt = (): void => {
-    if (useInputRow) {
-      if (rl !== null) {
-        statusBar.setInput(rl.line, rl.cursor);
-      }
-
-      statusBar.update(statusInfo());
-
-      return;
-    }
-
-    if (statusBar.active) {
-      statusBar.update(statusInfo());
-      process.stdout.write("\n› ");
-
-      return;
-    }
-
-    process.stdout.write("\n");
-    process.stdout.write(renderStatus(statusInfo()));
-    process.stdout.write("› ");
-  };
-
-  await new Promise<void>((resolveLoop) => {
-    let editorHandle: IEditorHandle | null = null;
-    let busy = false;
-    let closed = false;
-    let paletteOpen = false;
-
-    // Finish the loop only when stdin has closed AND no run is in flight — so a
-    // stdin EOF (piped input / Ctrl-D) never kills a build mid-turn.
-    const maybeFinish = (): void => {
-      if (closed && !busy) {
-        resolveLoop();
-      }
-    };
-
-    // Submit a line of input: check if busy/pending, echo it, handle /exit, or run it.
-    const submitLine = (raw: string): void => {
-      const line = raw.trim();
-
-      if (line.length === 0) {
-        if (!busy) {
-          prompt();
-        }
-
-        return;
-      }
-
-      // readline's output is sinked in input-row mode, so the submitted line is
-      // never echoed to scrollback — record it ourselves so the transcript reads
-      // naturally above the (now-cleared) input row.
-      if (useInputRow) {
-        echo(`\n${userBubble(line, true, process.stdout.columns)}\n`);
-      }
-
-      if (busy) {
-        if (line === "/exit" || line === "/quit") {
-          active?.abort();
-
-          if (rl !== null) {
-            rl.close();
-          }
-
-          if (editorHandle !== null) {
-            editorHandle.close();
-          }
-        } else {
-          pending.push(line);
-          echo("  ↳ queued (steers the next turn)\n");
-        }
-
-        return;
-      }
-
-      void runLine(line);
-    };
-
-    // Handle one idle line (slash command or a message), then any queued follow-up.
-    const runLine = async (line: string): Promise<void> => {
-      busy = true;
-      beginAgentTurn(); // the agent's response opens a fresh "▌ <model>" block
-
-      try {
-        if (line.startsWith("/")) {
-          if (await command(line)) {
-            if (rl !== null) {
-              rl.close();
-            }
-
-            return;
-          }
-        } else {
-          await dispatch(line);
-        }
-      } catch (err) {
-        // A command/turn that throws (e.g. a provider error mid-/compact) must NOT
-        // escape: runLine is invoked fire-and-forget (`void runLine(...)`), so an
-        // unhandled rejection would terminate the whole REPL — which read as "the
-        // CLI just exits". Surface the error and fall through to re-prompt instead.
-        spinner.stop(); // belt-and-suspenders: clear any spinner the failed path left running
-        echo(`\n⚠ ${err instanceof Error ? err.message : String(err)}\n`);
-      } finally {
-        closeAgentTurn(); // seal the agent card's bottom cap before re-prompting
-        busy = false;
-      }
-
-      // A line typed in the gap after the last steer-drain becomes the next turn.
-      const next = pending.shift();
-
-      if (next !== undefined) {
-        void runLine(next);
-
-        return;
-      }
-
-      if (closed) {
-        maybeFinish();
-      } else {
-        prompt();
-      }
-    };
-
-    // `/help` — the capability browser. On a TTY, opens an inline dropdown menu;
-    // off-TTY, prints the static help text so pipes/logs are unchanged. Extracted
-    // to keep cognitive complexity in check.
-    const buildHelpDeps = async (): Promise<
-      Parameters<typeof runCapabilityMenu>[0]
-    > => {
-      const suspend = (): void => {
-        editorControl?.suspend();
-        editorControl?.setInputInert(true);
-      };
-
-      const resume = (): void => {
-        editorControl?.setInputInert(false);
-        editorControl?.resume();
-        editorControl?.getBuffer().setText("");
-      };
-
-      const hasRecipes = (await loadRecipes(args.dir)).length > 0;
-
-      return {
-        color: process.stdout.isTTY,
-        hasRecipes,
-        runCommand: (c) => {
-          // c already includes the leading slash (registry stores "/sessions").
-          void runLine(c);
-        },
-        prefill: (c) => {
-          editorControl?.getBuffer().setText(`${c} `);
-        },
-        openWizard: async (opener) =>
-          opener === "scaffold"
-            ? openScaffoldInRepl({
-                cwd: args.dir,
-                suspend,
-                resume,
-                out: (s) => process.stdout.write(s),
-              })
-            : openRecipePicker({
-                cwd: args.dir,
-                render: (lines) => {
-                  statusBar.setOverlay(lines, statusInfo());
-                },
-                close: () => {
-                  statusBar.clearOverlay(statusInfo());
-                },
-                out: (s) => process.stdout.write(s),
-                runRecipe: (recipe) => {
-                  if (recipe.gate !== undefined) {
-                    session.setGate(recipe.gate);
-                    gateLabel = recipe.gate;
-                  }
-
-                  if (recipe.files !== undefined) {
-                    session.setScope([...recipe.files]);
-                  }
-
-                  if (recipe.task !== undefined) {
-                    void runLine(recipe.task);
-                  }
-                },
-              }),
-        render: (lines) => {
-          statusBar.setOverlay(lines, statusInfo());
-        },
-        close: () => {
-          statusBar.clearOverlay(statusInfo());
-        },
-      };
-    };
-
-    handleHelp = async (): Promise<void> => {
-      if (!process.stdout.isTTY) {
-        process.stdout.write(`${HELP}\n`);
-
-        return;
-      }
-
-      editorControl?.suspend();
-      editorControl?.setInputInert(true);
-
-      try {
-        const deps = await buildHelpDeps();
-
-        await runCapabilityMenu(deps);
-      } finally {
-        editorControl?.setInputInert(false);
-        editorControl?.resume();
-        editorControl?.getBuffer().setText("");
-      }
-
-      if (statusBar.active) {
-        statusBar.update(statusInfo());
-      }
-    };
-
-    // Helper: repaint the editor buffer to the status bar after palette insertion.
-    const repaintEditor = (handle: IEditorHandle): void => {
-      const { line, col } = handle.getBuffer().getCursor();
-      const lines = handle.getBuffer().getText().split("\n");
-
-      const frame = renderEditor(
-        {
-          lines,
-          cursorLine: line,
-          cursorCol: col,
-        },
-        {
-          columns: process.stdout.columns,
-          // Mirror the editor controller's own repaint window (rows minus the bar
-          // block) so wrapping/windowing matches.
-          maxRows: Math.max(1, process.stdout.rows - 3),
-          color: true,
-        }
-      );
-
-      // Repaint the editor block IN the pinned live region (setEditor), NOT via
-      // writeStream — writeStream treats its argument as conversation content, so
-      // it would strand the editor frame in scrollback (a leftover "/" per palette
-      // open). This mirrors the editor's renderEditor→setEditor callback.
-      statusBar.setEditor(
-        frame.frame.split("\n"),
-        frame.cursorRow,
-        frame.cursorCol
-      );
-    };
-
-    // Open the interactive `/` command palette: pick a command from a navigable
-    // list, then either run it (no-arg) or prefill the line so the user types the
-    // argument. Cancel ⇒ back to a clean prompt. Only meaningful on a TTY.
-    const openPalette = async (): Promise<void> => {
-      paletteOpen = true;
-      // Suspend the editor's stdin ownership so the palette's keypress loop owns
-      // input (see openFilePicker). Resumed in finally.
-      editorHandle?.suspend();
-
-      // Inline palette: paint the command list as an overlay above the input row
-      // (no alt-screen), same mechanism as the `@` picker and /help. The live
-      // query rides in the overlay title.
-      const view: IPaletteView = {
-        render: (lines) => {
-          statusBar.setOverlay(lines, statusInfo());
-        },
-        close: () => {
-          statusBar.clearOverlay(statusInfo());
-        },
-      };
-
-      try {
-        const picked = await pickCommand(view);
-
-        if (picked !== null) {
-          if (editorHandle !== null) {
-            editorHandle.getBuffer().setText("");
-
-            if (takesArg(picked)) {
-              // Prefill "<cmd> " so the user types the argument next.
-              editorHandle.getBuffer().insert(`${picked.name} `);
-              repaintEditor(editorHandle);
-            } else {
-              // No-arg command: run it and leave the input EMPTY. Inserting the
-              // name would linger in the buffer and reappear on the next keystroke
-              // (the "/clear" ghost after the screen is cleared).
-              repaintEditor(editorHandle);
-              void runLine(picked.name);
-            }
-          } else if (rl !== null) {
-            rl.write(null, { ctrl: true, name: "u" }); // clear the typed "/"
-
-            if (takesArg(picked)) {
-              rl.write(`${picked.name} `);
-            } else {
-              void runLine(picked.name);
-            }
-          }
-        } else if (editorHandle !== null) {
-          // Cancel (Esc / backspace-past-empty): drop the lingering trigger "/"
-          // so it doesn't stay in the input.
-          editorHandle.getBuffer().setText("");
-          repaintEditor(editorHandle);
-        } else if (rl !== null) {
-          rl.write(null, { ctrl: true, name: "u" });
-        }
-      } finally {
-        paletteOpen = false;
-
-        // Hand stdin back to the editor and repaint its input row (the overlay
-        // cleared it). No-op in readline mode (editorHandle is null).
-        if (editorHandle !== null) {
-          editorHandle.resume();
-          repaintEditor(editorHandle);
-        }
-
-        if (useInputRow) {
-          statusBar.update(statusInfo());
-
-          if (rl !== null) {
-            syncInput();
-          }
-        }
-      }
-    };
-
-    // Open the interactive `@` file picker: a compact dropdown rendered INLINE just
-    // above the input row (the conversation stays visible — no alternate screen),
-    // recency-ordered, type to fuzzy-filter. The buffer keeps its `@`; the live
-    // query is echoed onto the input row for feedback (it isn't in readline's/editor's
-    // buffer — the picker owns input). On select, the full path is appended after
-    // the `@`; at send time `@path` expands to the file's contents (see runSend).
-    const openFilePicker = async (): Promise<void> => {
-      paletteOpen = true;
-      // In editor mode the editor owns stdin via a `data` listener; suspend it so
-      // the inline picker's own `keypress` loop isn't fighting the editor for every
-      // keystroke (both would otherwise consume the same input). Resumed in finally.
-      editorHandle?.suspend();
-
-      const base =
-        editorHandle !== null
-          ? editorHandle.getBuffer().getText()
-          : rl !== null
-            ? rl.line
-            : ""; // text up to and including the just-typed `@`
-
-      const view: IPickerView = {
-        render: (query, items, selected): void => {
-          const rows = formatCompletionRows(
-            items,
-            selected,
-            process.stdout.columns,
-            process.stdout.isTTY
-          );
-
-          statusBar.setInput(`${base}${query}`, base.length + query.length);
-          statusBar.setOverlay(rows, statusInfo());
-        },
-        close: (): void => {
-          statusBar.clearOverlay(statusInfo());
-        },
-      };
-
-      try {
-        const files = await listWorkspaceFiles(args.dir);
-        const picked = await pickFileInline(files, view);
-
-        if (picked !== null) {
-          if (editorHandle !== null) {
-            editorHandle.getBuffer().insert(`${picked} `);
-            repaintEditor(editorHandle);
-          } else if (rl !== null) {
-            rl.write(`${picked} `);
-          }
-        }
-      } finally {
-        paletteOpen = false;
-
-        // Hand stdin back to the editor and repaint its input row (the overlay
-        // cleared it). No-op in readline mode (editorHandle is null).
-        if (editorHandle !== null) {
-          editorHandle.resume();
-          repaintEditor(editorHandle);
-        }
-
-        if (useInputRow) {
-          statusBar.update(statusInfo());
-
-          if (rl !== null) {
-            syncInput();
-          }
-        }
-      }
-    };
-
-    // `/` on an empty line opens the palette; `@` at a word boundary opens the file
-    // picker. The editor handles these internally (via openPalette/openFilePicker deps);
-    // readline mode uses keypress detection. The shared paletteOpen guard keeps the
-    // two overlays mutually exclusive. No-op while busy.
-
-    if (process.stdin.isTTY && !useEditor && !flags.basicInput()) {
-      // Only set up keypress detection for readline mode (not editor mode).
-      emitKeypressEvents(process.stdin);
-      process.stdin.on("keypress", (str: string | undefined) => {
-        syncInput(); // keep the pinned input row in sync as the user types
-
-        if (busy || paletteOpen) {
-          return;
-        }
-
-        if (str === "/" && rl !== null) {
-          setImmediate(() => {
-            if (!busy && !paletteOpen && rl.line === "/") {
-              void openPalette();
-            }
-          });
-        } else if (str === "@" && useInputRow && rl !== null) {
-          // The inline dropdown renders above the input row, so it needs that row
-          // (a tall-enough TTY). Without it we skip the picker — `@path` typed by
-          // hand still expands at send time (composeMessage), just no live popup.
-          setImmediate(() => {
-            if (
-              !busy &&
-              !paletteOpen &&
-              shouldOpenAtPicker(rl.line, rl.cursor)
-            ) {
-              void openFilePicker();
-            }
-          });
-        }
-      });
-    }
-
-    // Event-driven (not for-await) so stdin is read DURING a run: a line typed
-    // mid-run is queued to steer the next turn (or, if "/exit", aborts). This is
-    // what makes it feel like a real harness — you can redirect without waiting.
-    // When the editor is active, submitLine is wired via onSubmit; otherwise it's
-    // called here from readline. Crucially: the editor owns stdin exclusively in
-    // editor mode, and readline is NOT created in that case.
-    if (useEditor) {
-      // Editor-native `@`-completion: preload the workspace file list once, then
-      // filter it synchronously as the user types. The dropdown is painted ABOVE
-      // the editor block (not the readline input row), so it can't fight the editor
-      // for the cursor — the cause of the earlier display corruption.
-      let completionFiles: readonly string[] = [];
-
-      void listWorkspaceFiles(args.dir).then((files) => {
-        completionFiles = files;
-      });
-
-      const editorCompletion = {
-        items: (query: string): readonly string[] =>
-          filterFiles(completionFiles, query),
-        render: (items: readonly string[], selected: number): void => {
-          statusBar.setEditorOverlay(
-            formatCompletionRows(
-              items,
-              selected,
-              process.stdout.columns,
-              process.stdout.isTTY
-            )
-          );
-        },
-        clear: (): void => {
-          statusBar.clearEditorOverlay();
-        },
-      };
-
-      editorHandle = startEditor({
-        stdin: {
-          on: (event: string, cb: (data: string) => void) => {
-            process.stdin.on(event, cb);
-          },
-          removeListener: (event: string, cb: (data: string) => void) => {
-            process.stdin.removeListener(event, cb);
-          },
-          setRawMode: (mode: boolean) => {
-            process.stdin.setRawMode(mode);
-          },
-          resume: () => {
-            process.stdin.resume();
-          },
-          // The editor does string ops per chunk; without UTF-8 encoding,
-          // process.stdin emits Buffers and the first keypress crashes.
-          setEncoding: () => {
-            process.stdin.setEncoding("utf8");
-          },
-        },
-        out: (s: string) => {
-          statusBar.writeStream(s);
-        },
-        // Multi-row editor rendering callback: paints to the pinned input area
-        renderEditor: (
-          lines: string[],
-          cursorRow: number,
-          cursorCol: number
-        ) => {
-          statusBar.setEditor(lines, cursorRow, cursorCol);
-        },
-        // Reserve the `› ` prompt gutter the StatusBar paints in front of the
-        // editor block, so wrapping matches the visible width and the prompt row
-        // never exceeds `columns`.
-        columns: Math.max(1, process.stdout.columns - PROMPT_COLS),
-        rows: process.stdout.rows,
-        openPalette,
-        openFilePicker,
-        completion: editorCompletion,
-      });
-
-      resizeEditor = (columns, rows): void => {
-        editorHandle?.resize(Math.max(1, columns - PROMPT_COLS), rows);
-      };
-
-      editorControl = editorHandle;
-
-      editorHandle.onSubmit(submitLine);
-      editorHandle.onInterrupt(() => {
-        if (active === null) {
-          closed = true;
-          editorHandle?.close();
-          maybeFinish();
-        } else {
-          active.abort();
-        }
-      });
-      editorHandle.onExit(() => {
-        closed = true;
-        editorHandle?.close();
-        maybeFinish();
-      });
-      // Shift+Tab cycles the interactive mode (plan → normal → …).
-      editorHandle.onCycleMode(() => {
-        setMode(nextMode(currentModeId).id);
-      });
-    } else if (rl !== null) {
-      rl.on("line", submitLine);
-    }
-
-    rl?.on("close", () => {
-      closed = true;
-      editorHandle?.close();
-      statusBar.teardown();
-      maybeFinish();
-    });
-
-    // Pin the bar before the first turn so it's visible while that turn streams.
-    statusBar.install(statusInfo());
-
-    if (args.task.length > 0) {
-      void runLine(args.task); // sent as the first message; prompts when done
-    } else {
-      prompt();
-    }
-  });
-
-  statusBar.teardown(); // belt-and-suspenders: restore the terminal on loop exit
-  process.stdout.off("resize", handleResize); // don't pin the REPL closure
-  interactiveStream = null; // later/headless writes go straight to stdout again
-
-  return 0;
-}
-
-/** `/map [status|forget]` (REPL) and `tsforge map` — build/inspect the workspace
- *  map. The built map primes future sessions (and a `/clear`). */
-async function runMapCommand(dir: string, sub: string): Promise<void> {
-  if (sub === "status") {
-    process.stdout.write(`${await mapStatus(dir)}\n`);
-
-    return;
-  }
-
-  if (sub === "forget") {
-    const had = await forgetMap(dir);
-
-    process.stdout.write(
-      had ? "workspace map deleted\n" : "no map to delete\n"
-    );
-
-    return;
-  }
-
-  if (sub.length > 0) {
-    process.stdout.write(
-      `unknown map subcommand: ${sub} (use 'status', 'forget', or nothing to build)\n`
-    );
-
-    return;
-  }
-
-  process.stdout.write("building workspace map…\n");
-
-  try {
-    const map = await buildAndPersistMap(dir);
-
-    process.stdout.write(
-      map === null
-        ? "no tsconfig.json — nothing to map (the map is for TypeScript projects)\n"
-        : `mapped ${map.meta.totalFiles} files, ${map.hubs.length} hubs. Primes new sessions (/clear to apply now).\n`
-    );
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-
-    process.stdout.write(`map failed: ${message}\n`);
-  }
-}
-
-/** `/review` in the REPL — review the current change and print findings. */
-async function runReviewCommand(
-  provider: OpenAICompatibleProvider,
-  dir: string,
-  base: string
-): Promise<void> {
-  process.stdout.write("reviewing the current change…\n");
-
-  // Guard the REPL: a review error (git/fs/model) must not crash the session.
-  try {
-    const report = await reviewChange(provider, dir, {
-      ...(base.length > 0 ? { base } : {}),
-      log: (m) => process.stdout.write(`  ↳ ${m}\n`),
-    });
-
-    process.stdout.write(`\n${formatReport(report)}\n`);
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-
-    process.stdout.write(`\nreview failed: ${message}\n`);
-  }
 }
 
 /** Run the auto/explicit gate ONCE and return its distinct failing rule ids, so a
@@ -2562,69 +247,6 @@ async function applyRecipeArg(args: ICliArgs): Promise<number | null> {
   process.stdout.write(`using recipe '${recipe.id}'\n`);
 
   return null;
-}
-
-/** Resolve the newest `--log` JSONL under ~/.tsforge/logs, or "" if none. */
-async function newestLogFile(): Promise<string> {
-  try {
-    // Filenames are ISO-timestamp-prefixed, so lexicographic sort = chronological.
-    const names = (await readdir(logsDir()))
-      .filter((n) => n.endsWith(".jsonl"))
-      .sort();
-    const latest = names.at(-1);
-
-    return latest === undefined ? "" : join(logsDir(), latest);
-  } catch {
-    return "";
-  }
-}
-
-/** A user-supplied log path resolved against cwd, or "" when none was given. */
-function resolveLogArg(arg: string): string {
-  if (arg.length === 0) {
-    return "";
-  }
-
-  return isAbsolute(arg) ? arg : join(process.cwd(), arg);
-}
-
-/** `tsforge trace [logfile]` / `/trace` — summarize a `--log` run: model/tool
- *  calls, policy decisions (allow/ask/deny by risk), gate verdicts, and
- *  turns-to-green. Deterministic, no model call. With no path it prefers `prefer`
- *  (the live session log) and falls back to the newest log on disk. */
-async function runTraceCommand(arg: string, prefer = ""): Promise<number> {
-  let file = resolveLogArg(arg);
-
-  if (file.length === 0) {
-    file = prefer;
-  }
-
-  if (file.length === 0) {
-    file = await newestLogFile();
-  }
-
-  if (file.length === 0) {
-    process.stdout.write(
-      "no log to analyze — run with --log first, or pass a path\n"
-    );
-
-    return 1;
-  }
-
-  const text = await Bun.file(file)
-    .text()
-    .catch(() => "");
-  const events = parseEventLog(text);
-
-  if (events.length === 0) {
-    process.stdout.write(`no events parsed from ${file}\n`);
-
-    return 1;
-  }
-
-  process.stdout.write(`trace of ${file}\n\n${formatTrace(events)}\n`);
-
-  return 0;
 }
 
 async function traceMode(args: ICliArgs): Promise<number> {
